Store spriteCount[] in 8 bit (instead of 32)
There can maximally be 32 sprites on a line (normally only 4 or 8, but with
limitsprites=off openMSX can show all 32 sprites on one line). So we can store
the elements of the spriteCount[] array in 8 bits instead of 32.

(Un)fortunately the array was not that big, only 313 entries, so this change
doesn't result in a measurable performance difference.

+++ src/video/SpriteChecker.hh
@@ -7,6 +7,7 @@
 #include "DisplayMode.hh"
 #include "serialize_meta.hh"
 #include "unreachable.hh"
+#include <cstdint>
 
 namespace openmsx {
 
@@ -21,7 +22,7 @@ public:
 	  * If the sprite is less than 32 pixels wide,
 	  * the lower bits are unused.
 	  */
-	typedef unsigned SpritePattern;
+	typedef uint32_t SpritePattern;
 
 	/** Contains all the information to draw a line of a sprite.
 	  */
@@ -31,7 +32,7 @@ public:
 		SpritePattern pattern;
 		/** X-coordinate of sprite, corrected for early clock.
 		  */
-		short int x;
+		int16_t x;
 		/** Bit 3..0 are index in palette.
 		  * Bit 6 is 0 for sprite mode 1 like behaviour,
 		  * or 1 for OR-ing of sprite colors.
@@ -361,7 +362,7 @@ private:
 	  * In other words, spriteCount[i] is the number of sprites
 	  * in spriteBuffer[i].
 	  */
-	int spriteCount[313];
+	uint8_t spriteCount[313];
 
 	/** Is current display mode planar or not?
 	  * TODO: Introduce separate update methods for planar/nonplanar modes.
